# series-tracker/docs/swagger.yaml
definitions:
  models.Serie:
    properties:
      id:
        description: Unique identifier for the series
        type: integer
      lastEpisodeWatched:
        description: Last episode watched of the series
        type: integer
      ranking:
        description: Score of the series used for ranking
        type: integer
      status:
        description: Current status of the series; "Watching", "Plan to Watch", "Dropped",
          "Completed"
        type: string
      title:
        description: Title of the series
        type: string
      totalEpisodes:
        description: Quantity of episodes in the series
        type: integer
    type: object
info:
  contact: {}
paths:
  /api/series:
    get:
      consumes:
      - application/json
      description: Get a list of all series in the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Serie'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Retrieve all series
      tags:
      - series
    post:
      consumes:
      - application/json
      description: Inserts a new series into the database, make sure the series object
        includes all the necessary fields.
      parameters:
      - description: Series info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Serie'
      produces:
      - application/json
      responses:
        "201":
          description: Newly created series
          schema:
            $ref: '#/definitions/models.Serie'
        "400":
          description: Bad request, e.g, invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error, e.g, database error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new series
      tags:
      - series
  /api/series/{id}:
    delete:
      consumes:
      - application/json
      description: Inserts a new series into the database, make sure the series object
        includes all the necessary fields.
      parameters:
      - description: Series ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No content
        "400":
          description: Bad request, e.g, invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error, e.g, database error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove an existing series from the database
      tags:
      - series
    get:
      consumes:
      - application/json
      description: Get details of a series using the provided ID
      parameters:
      - description: Series ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Serie'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Retrieve a series by ID
      tags:
      - series
    put:
      consumes:
      - application/json
      description: Updates details of an existing series
      parameters:
      - description: Series ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Serie'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an existing series
      tags:
      - series
  /api/series/{id}/downvote:
    patch:
      consumes:
      - application/json
      description: Decreases the rating (score) of the series with the specified ID
        by one vote.
      parameters:
      - description: Series ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Series successfully downvoted
          schema:
            $ref: '#/definitions/models.Serie'
        "400":
          description: Invalid series ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Series not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Decrease series score
      tags:
      - series
  /api/series/{id}/episode:
    patch:
      consumes:
      - application/json
      description: Increments the current episode number of a series by one
      parameters:
      - description: Series ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated series status
          schema:
            $ref: '#/definitions/models.Serie'
        "400":
          description: Invalid input or status value
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Series not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Advance series episode count
      tags:
      - series
  /api/series/{id}/status:
    patch:
      consumes:
      - application/json
      description: Updates the status of the series with the specified ID.
      parameters:
      - description: Series ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated series status
          schema:
            $ref: '#/definitions/models.Serie'
        "400":
          description: Invalid input or status value
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Series not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update series status
      tags:
      - series
  /api/series/{id}/upvote:
    patch:
      consumes:
      - application/json
      description: Increases the rating (score) of the series with the specified ID
        by one vote.
      parameters:
      - description: Series ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Series successfully upvoted
          schema:
            $ref: '#/definitions/models.Serie'
        "400":
          description: Invalid series ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Series not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Increase series score
      tags:
      - series
swagger: "2.0"